Squashed commit of the following:
commit 38b3d1b510d79ea6ae80a43f6d671a3d8e3f408a
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 14:15:54 2025 +0700

    Update resume.pdf

commit 997533124cca51d9838987ae209527d8c22f0066
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 14:13:03 2025 +0700

    Update education.tsx

commit ac0421180555640d9f13d0def31e097c0c0dbe68
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 14:12:52 2025 +0700

    Update achievements.tsx

commit 9aaf4e1df8e7336b8adfb80a1ef625ae8973eae5
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 14:01:47 2025 +0700

    sort-by-date

commit e96a29f6141a101ed052d6446574a3e1689b6439
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 13:58:45 2025 +0700

    refactor

commit f0159673af3fb4f93b6638bf10021e497414becc
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 13:49:27 2025 +0700

    Update page.tsx

commit 0ed047a0e5ba5aae5bff59fdad23d45ff77cd023
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 13:45:47 2025 +0700

    fix: Languages

commit c8836ce53c50de9c5a19441ddabbbf5ea303c9a5
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 13:44:41 2025 +0700

    update: cert

commit 5b73c639c3480dd952b1194deb7e17ebdcdad6c4
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 13:41:20 2025 +0700

    update: exp

commit 01ce1f992cfb5aba24a3321cfd01b6f558d50340
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Aug 24 13:36:02 2025 +0700

    update: education

commit 5eccc87d5cee437e1dae905479db585ff0e71432
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Mon Jul 14 13:28:18 2025 +0700

    migrate

commit 44ad9103d8eb96638d91a19b578939a29cb8445a
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Mon Jul 14 13:25:16 2025 +0700

    restructure

commit 4e95b3cd5972911e446f022f7cb2712bb254a271
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Mon Jul 7 00:02:35 2025 +0700

    2025-07-07

commit e797b9e2575f1f348f1655c335a0e22c4ce8c238
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Sun Jul 6 23:35:36 2025 +0700

    2025-07-06-2335

commit 9a055813e7a9e0171c97fcde49406f29e22bc511
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Mon May 26 17:31:11 2025 +0700

    ver: 2025-05-06-am

commit 1af8690edcee25953f9cd438a13605e886745ae2
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Mon May 26 16:45:16 2025 +0700

    2025-05-26

commit 0e6f4651570b4459ac2256d41b616d1aac2a0a25
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Mon May 26 16:07:02 2025 +0700

    rm: blue-div-color

commit 0e488d878e8c511a65bd55702940b31ad5ed9516
Author: Pupipat Singkhorn <pupipat.sk@gmail.com>
Date:   Sat Apr 26 02:06:24 2025 +0700

    Add web resume info (#4)

    * Update resume.pdf

    * Update main.md

    * info update

    * feat: keywords-badge

    * fix: loading

commit 79972a3d3dfdc89e6036433f68b6efcbb588386e
Author: Pupipat Singkhorn <pupipat.sk@gmail.com>
Date:   Sat Apr 26 00:57:29 2025 +0700

    Resume (#3)

    * fix: dir-struct

    * 2025-04-26

    * fix: personal-web-link

    * update: 2025-04-26

commit 33a0d7536cf17a5c0e9b488f71621986efa2ee8e
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Fri Apr 25 23:36:11 2025 +0700

    Delete README.md

commit d25edee6b6bcba4ee6a954dd1314c7e1ac1b10a9
Author: pupipatsk <pupipat.sk@gmail.com>
Date:   Fri Apr 25 23:23:11 2025 +0700

    pull back deleted files
